day10/lab/requirements.txt
# Day 10 Lab — Data Pipeline & Observability
# pip install -r requirements.txt

python-dotenv>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.6.0
pyyaml>=6.0.1
pytest>=8.0.0
pydantic>=2